Refactor button and icon file structure
