--- src/components/elements/CloseIcon/CloseIcon.styled.js
@@ -1,4 +1,0 @@
-import styled from 'styled-components';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-export const CloseIconStyled = styled(FontAwesomeIcon)``;
