+++ src/components/pages/GenerationPage/GeneratePetitionModal/GeneratePetitionModal.styled.js
@@ -1,7 +1,5 @@
 import styled from 'styled-components';
 import { colorWhite } from '../../../../styles/colors';
-import { keyAndAmbientShadows } from '../../../../styles/shadows';
-import Button from '../../../elements/Button/Button';
 
 export const ModalContent = styled.div`
   position: relative;
@@ -18,24 +16,5 @@ export const ModalContent = styled.div`
 
   ul {
     margin: 1rem;
-  }
-`;
-
-export const CloseButton = styled(Button)`
-  position: absolute;
-  right: 0;
-  top: 0;
-  padding: 0 0.5rem;
-  border-radius: 0;
-  transition: none;
-  transform: none;
-
-  &:hover {
-    ${keyAndAmbientShadows.dp2};
-    transform: none;
-  }
-
-  &:active {
-    transform: none;
   }
 `;
